app/requirements.txt
pandas
numpy
fastapi[all]
python-bcb
DadosAbertosBrasil
gspread
gspread_dataframe
gspread_formatting
oauth2client